more locachoice cleanup.  moved the LCPlanUtils materials into the zzunused directory.  Overall, I can see that that might be more efficient, but with the current setup, we generate relatively few new plans and copy a lot, so the practical advantage would need to be evaluated.  As of now, I opt for simplicity.
