+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/frozenepsilons/DestinationSampler.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.bestresponse;
+package org.matsim.contrib.locationchoice.frozenepsilons;
 
 import org.matsim.contrib.locationchoice.DestinationChoiceConfigGroup;
