+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/frozenepsilons/DestinationScoring.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.bestresponse;
+package org.matsim.contrib.locationchoice.frozenepsilons;
 
 import java.util.Collection;
 import java.util.Random;
