+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/frozenepsilons/MaxDCScoreWrapper.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.locationchoice.bestresponse;
+package org.matsim.contrib.locationchoice.frozenepsilons;
 
 import org.matsim.utils.objectattributes.ObjectAttributes;
