+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/frozenepsilons/ReadOrCreateKVals.java
@@ -18,7 +18,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.bestresponse;
+package org.matsim.contrib.locationchoice.frozenepsilons;
 
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Scenario;
