+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/zzunused/DestinationChoiceInitializer.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.locationchoice.zzunused;
 
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.TransportMode;
-import org.matsim.contrib.locationchoice.bestresponse.DestinationChoiceContext;
+import org.matsim.contrib.locationchoice.frozenepsilons.DestinationChoiceContext;
 import org.matsim.core.controler.MatsimServices;
 import org.matsim.core.controler.events.StartupEvent;
 import org.matsim.core.controler.listener.StartupListener;
