+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/zzunused/LCActivity.java
@@ -18,7 +18,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.bestresponse;
+package org.matsim.contrib.locationchoice.zzunused;
 
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
