+++ contribs/locationchoice/src/test/java/org/matsim/contrib/locationchoice/FrozenEpsilonLocaChoiceIT.java
@@ -16,10 +16,10 @@ import org.matsim.api.core.v01.network.Node;
 import org.matsim.api.core.v01.population.*;
 import org.matsim.contrib.analysis.kai.KaiAnalysisListener;
 import org.matsim.contrib.locationchoice.DestinationChoiceConfigGroup.ApproximationLevel;
-import org.matsim.contrib.locationchoice.bestresponse.BestReplyLocationChoicePlanStrategy;
-import org.matsim.contrib.locationchoice.bestresponse.DCActivityWOFacilitiesScoringFunction;
-import org.matsim.contrib.locationchoice.bestresponse.DCScoringFunctionFactory;
-import org.matsim.contrib.locationchoice.bestresponse.DestinationChoiceContext;
+import org.matsim.contrib.locationchoice.frozenepsilons.BestReplyLocationChoicePlanStrategy;
+import org.matsim.contrib.locationchoice.frozenepsilons.DCActivityWOFacilitiesScoringFunction;
+import org.matsim.contrib.locationchoice.frozenepsilons.DCScoringFunctionFactory;
+import org.matsim.contrib.locationchoice.frozenepsilons.DestinationChoiceContext;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.config.groups.PlanCalcScoreConfigGroup;
@@ -31,7 +31,6 @@ import org.matsim.core.controler.events.ShutdownEvent;
 import org.matsim.core.controler.listener.ShutdownListener;
 import org.matsim.core.gbl.Gbl;
 import org.matsim.core.population.PopulationUtils;
-import org.matsim.core.replanning.strategies.DefaultPlanStrategiesModule;
 import org.matsim.core.replanning.strategies.DefaultPlanStrategiesModule.DefaultSelector;
 import org.matsim.core.router.TripRouter;
 import org.matsim.core.router.TripStructureUtils;
